+++ src/globus_sdk/experimental/tokenstorage/sqlite.py
@@ -6,7 +6,7 @@ import sqlite3
 import textwrap
 import typing as t
 
-from globus_sdk.experimental.tokenstorage_v2.base import FileTokenStorage
+from globus_sdk.experimental.tokenstorage.base import FileTokenStorage
 from globus_sdk.version import __version__
 
 from .token_data import TokenData
